requirements.txt
Flask>=2.3.0
python-nut2>=0.4.0
configparser>=5.0.0